move MP4 to the bottom and hide disabled formats
Currently, mp4 is one of the first formats chosen. But there are some issues when it comes to scrubbing/seeking. Therefore we only use mp4 as the last resort, this should give a better user experience.

sort of fixes #590 & #626 & #487

Also when choosing the AVPlayer in the Quality settings, disabled formats are hidden.