remove orientationDebouncer
The orientationdebouncer was always firing on rotation changes, interfering with lockOrientation calls from the player. This resulted in multiple rotations to the same orientation. Visible for the user were multiple changes in the video size when rotating the device.

Also when Browser was locked to portrait, the very first landscape rotation did not register and needed to be called twice.