changes to Rotation Lock button in player
The rotation lock button function now depends on the Settings.

When "Follow device orientation lock" is enabled. Enabling rotation lock in the player will not rotate into landscape when hitting the fullscreen button.

When "Follow device orientation lock" is disabled, it will prevent the player from going into fullscreen on device rotation, but goes into fullscreen with landscape when hitting the fullscreen button.